Reorganized the code such that adding and checking ad format is now more modular. The ads are now in an array in one place, so there is no need to update several places in the code any time we add them.
Note however that not all ad formats are supported, and duplicate dimensions
are not allowed, such as image ads and text ads of the same dimension, or
text ads and ad links of the same dimension.

This is kind of an artificial restriction, but we have to live with it for
the time being until we properly implement it in the future.
Perhaps the array key should be the entire ad identifier (dimension + suffix),
and that needs to be referenced in the calls from the theme and/or block.